flow_cytometry_microservice/requirements.txt
anytree==2.12.1
bcrypt==4.2.0
certifi==2024.8.30
cryptography==43.0.3
et_xmlfile==2.0.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-PyMongo==2.3.0
FlowIO==1.3.0
FlowKit==1.2.3
FlowUtils==1.1.0
lxml==5.3.0
numpy==1.26.4
openpyxl==3.1.5
pandas==2.2.3
pathspec==0.12.1
PyJWT==2.9.0
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
setuptools==75.8.0
tzdata==2024.2
urllib3==2.2.3
watchdog==5.0.3
Werkzeug==3.0.4
